requirements.txt
fastapi[standard]
sqlmodel
pydantic-settings
passlib[bcrypt] 
pyjwt